cleanup

+++ cpp/includes/tidesurf/math_util.h
@@ -2,11 +2,8 @@
 #include <cstdint>
 #include "tidesurf/tidesurf_macros.h"
 namespace tidesurf {
-
-uint32_t power(uint32_t base, uint32_t exp);
-
-uint32_t num_10th(int64_t num);
-
+    uint32_t power(uint32_t base, uint32_t exp);
+    uint32_t num_10th(uint32_t num);
 }
